chore: move createTag to block-helpers.js
