+++ blocks/block-helpers.js
@@ -18,18 +18,25 @@ export const replaceElementType = (el, type) => {
 };
 
 /**
- * Create element with attributes
- * @param {string} name The tag of the element, ex) <div>
- * @param {object} attrs The attributes to add to the element, ex) {class: 'class-name'}
- * @returns el Created Element
+ * Create an element with ID, class, children, and attributes
+ * @param {String} tag the tag nav of the element
+ * @param {Object} attributes the attributes of the tag
+ * @param {HTMLElement} html the content of the element
+ * @returns {HTMLElement} the element created
  */
-export function createTag(name, attrs) {
-  const el = document.createElement(name);
-  if (typeof attrs === 'object') {
-    // eslint-disable-next-line no-restricted-syntax
-    for (const [key, value] of Object.entries(attrs)) {
-      el.setAttribute(key, value);
+export function createTag(tag, attributes, html) {
+  const el = document.createElement(tag);
+  if (html) {
+    if (html instanceof HTMLElement) {
+      el.append(html);
+    } else {
+      el.insertAdjacentHTML('beforeend', html);
     }
+  }
+  if (attributes) {
+    Object.keys(attributes).forEach((key) => {
+      el.setAttribute(key, attributes[key]);
+    });
   }
   return el;
 }
