+++ blocks/footer/footer.js
@@ -3,7 +3,7 @@ import {
   debug,
   makeLinksRelative,
 } from '../../scripts/scripts.js';
-import createTag from '../gnav/gnav-utils.js';
+import { createTag } from '../block-helpers.js';
 
 const GLOBE_IMG = '<img class="footer-region-img" loading="lazy" src="/blocks/footer/globe.svg">';
 const ADCHOICE_IMG = '<img class="footer-link-img" loading="lazy" src="/blocks/footer/adchoices-small.svg">';
