--- blocks/gnav/gnav-utils.js
@@ -1,23 +1,0 @@
-/**
- * Create an element with ID, class, children, and attributes
- * @param {String} tag the tag nav of the element
- * @param {Object} attributes the attributes of the tag
- * @param {HTMLElement} html the content of the element
- * @returns {HTMLElement} the element created
- */
-export default function createTag(tag, attributes, html) {
-  const el = document.createElement(tag);
-  if (html) {
-    if (html instanceof HTMLElement) {
-      el.append(html);
-    } else {
-      el.insertAdjacentHTML('beforeend', html);
-    }
-  }
-  if (attributes) {
-    Object.keys(attributes).forEach((key) => {
-      el.setAttribute(key, attributes[key]);
-    });
-  }
-  return el;
-}
